// lib/types/skill.ts
export interface Skill {
  id: number;
  name: string;
  slug: string;
  dark_icon_url: string;
  light_icon_url: string;
}

export interface SkillFormData {
  id: number;
  name: string;
  dark_icon_url: string;
  light_icon_url: string;
}

export interface SkillEditData {
  id: number;
  name: string;
  dark_icon_url: string;
  light_icon_url: string;
}